[OMEMO]: Attempt to fix bundle publishing
Use the following options in `omemo_bundle_publish()`:

- "pubsub#persist_items" = "true"
- "pubsub#access_model" = "open"

The same options are also used in Gajim.

Note that the XEP [1] also requires the option "pubsub#max_items" = "max",
but this seems to not work and is also not used in Gajim.

[1]: <https://xmpp.org/extensions/xep-0384.html#bundles>

I've tested this on two different servers. The bundle was successfully
added as a new PEP node.

TODO: The function `_omemo_bundle_publish_result` fails. Perhaps it's
not needed any more, as the relevant options are set in the stanza.